+++ include/quill/core/LoggerManager.h
@@ -151,6 +151,19 @@ public:
   }
 
   /***/
+  template <typename TLogger>
+  LoggerBase* create_or_get_logger(std::string const& logger_name, LoggerBase* source_logger)
+  {
+    if (!source_logger)
+    {
+      return get_logger(logger_name);
+    }
+
+    return create_or_get_logger<TLogger>(logger_name, source_logger->sinks, source_logger->pattern_formatter_options,
+                                         source_logger->clock_source, source_logger->user_clock);
+  }
+
+  /***/
   void remove_logger(LoggerBase* logger)
   {
     logger->mark_invalid();